+++ packages/editor/editor-core/src/__tests__/integration/emoji/emoji-2.ts
@@ -1,21 +1,19 @@
 import { BrowserTestCase } from '@atlaskit/webdriver-runner/runner';
 import Page from '@atlaskit/webdriver-runner/wd-wrapper';
-import { getDocFromElement } from '../_helpers';
 import {
-  messageEditor,
+  getDocFromElement,
   editable,
-  insertEmoji,
-  emojiItem,
-  insertEmojiBySelect,
-} from './_emoji-helpers';
-import { messages } from '../../../plugins/insert-block/ui/ToolbarInsertBlock';
+  gotoEditor,
+  insertBlockMenuItem,
+} from '../_helpers';
+import { insertEmoji, emojiItem, insertEmojiBySelect } from './_emoji-helpers';
 
 BrowserTestCase(
   'emoji-2.ts: should be able to use emoji inside blockquote',
   { skip: ['ie'] },
   async client => {
     const browser = new Page(client);
-    await browser.goto(messageEditor);
+    await gotoEditor(browser);
     await browser.type(editable, '> ');
     await browser.type(editable, 'some text ');
     await insertEmoji(browser, 'a');
@@ -30,7 +28,7 @@ BrowserTestCase(
   { skip: ['ie'] },
   async client => {
     const browser = new Page(client);
-    await browser.goto(messageEditor);
+    await gotoEditor(browser);
     await browser.type(editable, '* ');
     await insertEmoji(browser, 'smile');
     await browser.waitForSelector(emojiItem('smile'));
@@ -44,7 +42,7 @@ BrowserTestCase(
   { skip: ['ie'] },
   async client => {
     const browser = new Page(client);
-    await browser.goto(messageEditor);
+    await gotoEditor(browser);
     await browser.type(editable, '1. ');
     await insertEmoji(browser, 'a');
     await browser.waitForSelector(emojiItem('a'));
@@ -56,11 +54,10 @@ BrowserTestCase(
 // ie keying in ; instead of : - browserstack issue
 BrowserTestCase(
   'emoji-2.ts: should be able remove emoji on backspace',
-  { skip: ['safari', 'ie'] },
+  { skip: ['safari', 'ie', 'firefox', 'edge'] },
   async client => {
     const browser = new Page(client);
-    await browser.goto(messageEditor);
-    await browser.waitForSelector(editable);
+    await gotoEditor(browser);
     await browser.type(editable, 'this ');
     await insertEmoji(browser, 'joy');
     await browser.waitForSelector(emojiItem('joy'));
@@ -78,8 +75,7 @@ BrowserTestCase(
   async client => {
     const decisions = 'span[aria-label="Decision"]';
     const browser = new Page(client);
-    await browser.goto(messageEditor);
-    await browser.waitForSelector(editable);
+    await gotoEditor(browser);
     // to get steps working on edge since its is slow
     await browser.type(editable, '<> ');
     await browser.waitForSelector(decisions);
@@ -95,16 +91,11 @@ BrowserTestCase(
   { skip: ['ie'] },
   async client => {
     const decisions = 'li span[aria-label="Decision"]';
-    const createDecisions = `span[aria-label="${
-      messages.decision.defaultMessage
-    }"]`;
     const browser = new Page(client);
-    await browser.goto(messageEditor);
-    await browser.waitForSelector(editable);
+    await gotoEditor(browser);
     await browser.type(editable, 'this ');
     await insertEmoji(browser, 'smile');
-    await browser.waitForSelector(createDecisions);
-    await browser.click(createDecisions);
+    await insertBlockMenuItem(browser, 'Decision');
     await browser.waitForSelector(decisions);
     await browser.isExisting(decisions);
     const doc = await browser.$eval(editable, getDocFromElement);